Add back Bad ratings to negatively affect overall ratings
If the player makes a hit that they already got OK or Perfect for or are
wildly out of time, the Bad hit counter will be incremented. This is
necessary because re-making a hit can happen any number of times and
cannot be determined from the other hit rating counts.