Move dummy instrument back to ROM0
Keeping it in ROMX means the instrument bank variables need to be
updated in Music_Pause or SFX_Stop, which they aren't. However, it's not
necessary to do that if it can just be in ROM0 (where it originally
was). bgb doesn't like bankswitches to bank 0, so intialize the bank
number to 1.