Don't use the include directory
It might be confusing to see `INCLDUE "actors.inc"` with all the other
INCLUDEs being files in the constants directory. The include path can be
removed from the Makefile, but seeing an include directory along with
other more specific directories might also be confusing.

Files in the include directory have been moved to the constants and the
(new) macros directory. SoundSystem files have been moved to their own
SoundSystem directory.