Medals for the main menu
the large medals would be used on the panel featuring a little description of the game when selecting it on the main menu. the small ones (which come to think of it should probably be in the background but I really don't think we are gonna run out of OAM on the main menu) would be next to a minigame so just at a glance the player can see which minigames they have a medal in and what kind of medal it is. the "new" and "!" tiles denote a minigame the player is yet to play, the blank medal corresponds to having no small medal on the minigame icon and represents the player's best score being bad. Bronze or dark grey is for ok, silver or light grey is for great, and gold or white is for perfect.