Use numerically correct values in next hit timer
Using the "increment high byte if low byte is non-zero" trick doesn't
work when actually trying to figure out how on-time the player is. For
example, if the next hit timer is $100, it's actually $200. Simply
adjusting doesn't work, since if the timer is $101, it's actually 1.
This trick is great for countdowns, but it's not as great when the value
of that countdown is actually needed.