Fix bugs with extra LYC interrupts
Handling so many interrupts of the same type at the same time is proving
difficult! It's almost fully functional now, though, and the only
problem left is with extra LYC interrupts on the same scanlines as
transition LYC interrupts, causing the transition block not to be
updated. A 1-scanline offset shouldn't be noticeable, and will likely
happen anyway due to both interrupt handlers waiting for HBlank before
messing with stuff, so it seems like an acceptable solution.