Move sound update calls to main loops
Having the sound update in an LYC interrupt was a little problematic. It
would delay some other LYC interrupts when returning, causing (albeit
barely noticeable) graphical issues. Updating sound in an interrupt
handler is good for preventing delays, since they fire regularly, but if
the game experiences a delay, it will actually be *desynced* with the
music. In a rhythm game, desynced music is... not good.

Unfortunately, removing the sound update interrupt means that sometimes
there will be no interrupts and interrupts should be disabled. This
complicates things a little, but it was doable.